pyproject.toml
[project]
name = "mscl-rs"
version = "0.1.0"
description = "Parser for MicroStrain Communication Library (MSCL) packets in Rust."
readme = "README.md"
authors = [
    { name = "Harshil", email = "37377066+harshil21@users.noreply.github.com" }
]
requires-python = ">=3.10"

[tool.uv]
# Rebuild package when any rust files change
cache-keys = [{file = "pyproject.toml"}, {file = "Cargo.toml"}, {file = "**/*.rs"}]
# Uncomment to build rust code in development mode (no optimizations, faster builds)
# config-settings = { build-args = '--profile=dev' }

[dependency-groups]
dev = [
    "pyserial",
    "msgspec>=0.20.0",
    "pytest",
    "python-mscl; python_version <= '3.13'",
    "maturin",
    "ruff",
    "ziglang"  # for maturin's zig support when cross-compiling
]

[build-system]
requires = ["maturin>=1.0"]
build-backend = "maturin"

# For uploading to TestPyPI
# [[tool.uv.index]]
# name = "testpypi"
# url = "https://test.pypi.org/simple/"
# publish-url = "https://test.pypi.org/legacy/"
# explicit = true
